backend/requirements.txt
# CodeRefine Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
google-generativeai>=0.4.0,<1.0
python-dotenv==1.0.1
